requirements.txt
alembic==1.14.0
anyio==4.7.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.12.14
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.2
rfc3986==1.5.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tornado==6.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.20
